new version

+++ ermaFilt.h
@@ -4,11 +4,11 @@
 extern IIRFILTER numerFilter, denomFilter;
 
 void ermaFiltPrep();
-void ermaDownsample(float *X,  int32_t nX,		/* in */
-		    long decim,				/* in */
-		    float *Y, int32_t *nY,		/* out */
+void ermaDownsample(float *X,  int32 nX,		/* in */
+		    int32 decim,			/* in */
+		    float *Y, int32 *nY,		/* out */
 		    float inSRate, float *outSRate);	/* in, out */
-void ermaNumerDenomFilt(float *X, long nX,		/* in */
+void ermaNumerDenomFilt(float *X, int32 nX,		/* in */
 			float *numer, float *denom);	/* out */
 void ermaFiltGetBandwidths(float *pNumerBW, float *pDenomBW);
