+++ wavFile.h
@@ -4,7 +4,7 @@
 /*#include "wisprFile.h"		*//* for WISPRINFO */
 
 WISPRINFO *wavReadHeader(WISPRINFO *wi, char *filename);
-int wavReadData(int16_t *buf, WISPRINFO *wi, long offset, long nSamp);
+int wavReadData(int16_t *buf, WISPRINFO *wi, size_t offset, size_t nSamp);
 void wavCloseFile(WISPRINFO *wi);
 double getTimeFromName(char *fn);
